Move annotations to separate file
